change dates selection for total coins stats, init cache concurrently (10)
