+++ tokenomics/balance.go
@@ -21,37 +21,6 @@ import (
 	"github.com/ice-blockchain/wintr/time"
 )
 
-func (r *repository) GetTotalCoinsSummary(ctx context.Context, days uint64, utcOffset stdlibtime.Duration) (*TotalCoinsSummary, error) {
-	var (
-		dates    []stdlibtime.Time
-		res      = &TotalCoinsSummary{TimeSeries: make([]*TotalCoinsTimeSeriesDataPoint, 0, days)}
-		now      = time.Now()
-		location = stdlibtime.FixedZone(utcOffset.String(), int(utcOffset.Seconds()))
-	)
-
-	dates, res.TimeSeries = r.totalCoinsDates(now, days, r.cfg.GlobalAggregationInterval.Parent)
-	totalCoins, err := r.getCachedTotalCoins(ctx, dates)
-	if err != nil {
-		return nil, errors.Wrapf(err, "failed to getCachedTotalCoins for createdAts:%#v", dates)
-	}
-	for _, child := range res.TimeSeries {
-		for _, stats := range totalCoins {
-			dayMatch := stats.CreatedAt.Equal(child.Date)
-			if dayMatch {
-				child.Standard = stats.BalanceTotalStandard
-				child.PreStaking = stats.BalanceTotalPreStaking
-				child.Blockchain = stats.BalanceTotalEthereum
-				child.Total = child.Standard + child.PreStaking + child.Blockchain
-				break
-			}
-		}
-		child.Date = child.Date.In(location)
-	}
-	res.TotalCoins = res.TimeSeries[0].TotalCoins
-
-	return res, nil
-}
-
 func (r *repository) GetBalanceSummary( //nolint:lll // .
 	ctx context.Context, userID string,
 ) (*BalanceSummary, error) {
@@ -336,130 +305,4 @@ func ApplyPreStaking(amount, preStakingAllocation, preStakingBonus float64) (flo
 	preStakingAmount := (amount * (100 + preStakingBonus) * preStakingAllocation) / 10000
 
 	return standardAmount, preStakingAmount
-}
-
-func (r *repository) totalCoinsDates(now *time.Time, days uint64, dayInterval stdlibtime.Duration) ([]stdlibtime.Time, []*TotalCoinsTimeSeriesDataPoint) {
-	var (
-		adjustForLatencyToProcessAllUsers = -(r.cfg.GlobalAggregationInterval.Child / 4)
-		truncationInterval                = r.cfg.GlobalAggregationInterval.Child
-		dates                             = make([]stdlibtime.Time, 0, days)
-		timeSeries                        = make([]*TotalCoinsTimeSeriesDataPoint, 0, days)
-	)
-	for day := uint64(0); day < days; day++ {
-		date := now.Add(dayInterval * -1 * stdlibtime.Duration(day)).Add(adjustForLatencyToProcessAllUsers).Truncate(truncationInterval)
-		dates = append(dates, date)
-		timeSeries = append(timeSeries, &TotalCoinsTimeSeriesDataPoint{Date: date})
-	}
-	return dates, timeSeries
-}
-
-func (r *repository) cacheTotalCoins(ctx context.Context, coins []*dwh.TotalCoins) error {
-	val := make([]interface{ Key() string }, 0, len(coins))
-	for _, v := range coins {
-		val = append(val, v)
-	}
-	return errors.Wrapf(storage.Set(ctx, r.db, val...), "failed to set cache value for total coins: %+v", coins)
-}
-
-func (r *repository) getCachedTotalCoins(ctx context.Context, dates []stdlibtime.Time) ([]*dwh.TotalCoins, error) {
-	keys := make([]string, 0, len(dates))
-	for _, d := range dates {
-		keys = append(keys, r.totalCoinsCacheKey(d))
-	}
-	cacheResult, err := storage.Get[dwh.TotalCoins](ctx, r.db, keys...)
-	if err != nil {
-		return nil, errors.Wrapf(err, "failed to get cached coinStats for dates %v", dates)
-	}
-	cached := make([]*dwh.TotalCoins, 0, len(cacheResult))
-	for _, c := range cacheResult {
-		cached = append(cached, &dwh.TotalCoins{
-			CreatedAt:              c.CreatedAt,
-			BalanceTotalStandard:   c.BalanceTotalStandard,
-			BalanceTotalPreStaking: c.BalanceTotalPreStaking,
-			BalanceTotalEthereum:   c.BalanceTotalEthereum,
-		})
-	}
-
-	return cached, nil
-}
-
-func (r *repository) totalCoinsCacheKey(date stdlibtime.Time) string {
-	return fmt.Sprintf("totalCoinStats:%v", date.Truncate(r.cfg.GlobalAggregationInterval.Child).Format(stdlibtime.RFC3339))
-}
-
-func (r *repository) keepTotalCoinsCacheUpdated(ctx context.Context) {
-	for ctx.Err() == nil {
-		now := time.Now()
-		if now.Sub(now.Truncate(r.cfg.GlobalAggregationInterval.Child)) >= stdlibtime.Duration(float64(r.cfg.GlobalAggregationInterval.Child)*0.75) {
-			dwhCtx, cancel := context.WithTimeout(ctx, 1*stdlibtime.Minute)
-			if err := r.buildTotalCoinCache(dwhCtx, now.Truncate(r.cfg.GlobalAggregationInterval.Child)); err != nil {
-				log.Error(errors.Wrapf(err, "failed to update total coin stats cache for date %v", *now.Time))
-			}
-			cancel()
-		} else {
-			stdlibtime.Sleep(stdlibtime.Second)
-		}
-	}
-}
-
-func (r *repository) buildTotalCoinCache(ctx context.Context, dates ...stdlibtime.Time) error {
-	totalCoins, err := r.dwh.SelectTotalCoins(ctx, dates)
-	if err != nil {
-		return errors.Wrapf(err, "failed to read total coin stats cacheable values for dates %v", dates)
-	}
-	for _, d := range dates {
-		found := false
-		for _, stats := range totalCoins {
-			if stats.CreatedAt.Equal(d) {
-				found = true
-				break
-			}
-		}
-		if !found {
-			totalCoins = append(totalCoins, &dwh.TotalCoins{
-				CreatedAt:              time.New(d),
-				BalanceTotalStandard:   0,
-				BalanceTotalPreStaking: 0,
-				BalanceTotalEthereum:   0,
-			})
-		}
-	}
-	return errors.Wrapf(
-		r.cacheTotalCoins(ctx, totalCoins),
-		"failed to save total coin stats cache for dates %v", dates)
-}
-
-func (r *repository) mustInitTotalCoinsCache(ctx context.Context) {
-	now := time.Now()
-	childIntervalsInParent := r.cfg.GlobalAggregationInterval.Parent / r.cfg.GlobalAggregationInterval.Child
-	dates, _ := r.totalCoinsDates(now, uint64(daysCountToInitCoinsCacheOnStartup*childIntervalsInParent), r.cfg.GlobalAggregationInterval.Child)
-	alreadyCached, err := r.getCachedTotalCoins(ctx, dates)
-	if err != nil {
-		log.Panic(errors.Wrapf(err, "failed to init total coin stats cache"))
-	}
-	for _, cached := range alreadyCached {
-		for dateIdx, date := range dates {
-			if cached.CreatedAt.Equal(date) {
-				dates = append(dates[:dateIdx], dates[dateIdx+1:]...)
-
-				break
-			}
-		}
-	}
-	datesSameDay := make([]stdlibtime.Time, 0, childIntervalsInParent)
-	for i := range dates {
-		if i > 0 && dates[i-1].Truncate(r.cfg.GlobalAggregationInterval.Parent).Equal(dates[i].Truncate(r.cfg.GlobalAggregationInterval.Parent)) {
-			datesSameDay = append(datesSameDay, dates[i-1])
-		} else if i > 0 && len(datesSameDay) > 0 {
-			datesSameDay = append(datesSameDay, dates[i-1])
-			log.Info(fmt.Sprintf("Building total coins cache %v...", datesSameDay[0]))
-			err = errors.Wrapf(r.buildTotalCoinCache(ctx, datesSameDay...), "failed to build/init total coins cache for %v", datesSameDay)
-			for err != nil {
-				log.Error(err)
-				err = errors.Wrapf(r.buildTotalCoinCache(ctx, datesSameDay...), "failed to build/init total coins cache for %v", datesSameDay)
-			}
-			datesSameDay = datesSameDay[:0]
-		}
-
-	}
 }
